+++ examples/quantized_net/tutorial_ternaryweight_cifar10_tfrecord.py
@@ -2,8 +2,8 @@
 # -*- coding: utf-8 -*-
 """
 
-- 1. This model has 1,068,298 paramters and Dorefa compression strategy(weight:1 bit, active: 3 bits),
-after 500 epoches' training with GPU,accurcy of 81.1% was found.
+- 1. This model has 1,068,298 paramters and TWN compression strategy(weight:1,0,-1, output: float32),
+after 500 epoches' training with GPU,accurcy of 80.6% was found.
 
 - 2. For simplified CNN layers see "Convolutional layer (Simplified)"
 in read the docs website.
@@ -12,8 +12,8 @@ in read the docs website.
 
 Links
 -------
-.. paper:https://arxiv.org/abs/1606.06160
-.. code:https://github.com/XJTUWYD/DoReFa_Cifar10
+.. https://arxiv.org/abs/1605.04711
+.. https://github.com/XJTUWYD/TWN
 
 Note
 ------
@@ -38,15 +38,15 @@ of processing time. To prevent these operations from slowing down training,
 we run them inside 16 separate threads which continuously fill a TensorFlow queue.
 
 """
-
 import multiprocessing
 import time
 
 import numpy as np
 import tensorflow as tf
 import tensorlayer as tl
-from tensorlayer.layers import (Conv2d, Dense, DorefaConv2d, DorefaDense,
-                                Flatten, Input, LocalResponseNorm, MaxPool2d)
+from tensorlayer.layers import (
+    Conv2d, Dense, Flatten, Input, LocalResponseNorm, MaxPool2d, TernaryConv2d, TernaryDense
+)
 from tensorlayer.models import Model
 
 tl.logging.set_verbosity(tl.logging.DEBUG)
@@ -56,24 +56,31 @@ tl.logging.set_verbosity(tl.logging.DEBUG)
 X_train, y_train, X_test, y_test = tl.files.load_cifar10_dataset(shape=(-1, 32, 32, 3), plotable=False)
 
 
-def dorefanet_model(input_shape, n_classes):
+def model(input_shape, n_classes):
     in_net = Input(shape=input_shape, name='input')
-    net = Conv2d(32, (5, 5), (1, 1), act='relu', padding='SAME', name='conv1')(in_net)
+
+    net = Conv2d(64, (5, 5), (1, 1), act=tf.nn.relu, padding='SAME', name='cnn1')(in_net)
     net = MaxPool2d((3, 3), (2, 2), padding='SAME', name='pool1')(net)
     net = LocalResponseNorm(4, 1.0, 0.001 / 9.0, 0.75, name='norm1')(net)
-    net = tl.layers.Sign("sign")(net)
-    net = DorefaConv2d(8, 32, 64, (5, 5), (1, 1), act='relu', padding='SAME', name='DorefaConv1')(net)
+
+    net = TernaryConv2d(64, (5, 5), (1, 1), act=tf.nn.relu, padding='SAME', name='cnn2')(net)
     net = LocalResponseNorm(4, 1.0, 0.001 / 9.0, 0.75, name='norm2')(net)
     net = MaxPool2d((3, 3), (2, 2), padding='SAME', name='pool2')(net)
+
     net = Flatten(name='flatten')(net)
-    net = DorefaDense(8, 16, 384, act='relu', name='DorefaDense1')(net)
-    net = DorefaDense(8, 16, 192, act='relu', name='DorefaDense2')(net)
+
+    net = TernaryDense(384, act=tf.nn.relu, name='d1relu')(net)
+    net = TernaryDense(192, act=tf.nn.relu, name='d2relu')(net)
     net = Dense(n_classes, act=None, name='output')(net)
+
     net = Model(inputs=in_net, outputs=net, name='dorefanet')
     return net
 
+
 # training settings
-net = dorefanet_model([None, 24, 24, 3], n_classes=10)
+bitW = 8
+bitA = 8
+net = model([None, 24, 24, 3], n_classes=10)
 batch_size = 128
 n_epoch = 50000
 learning_rate = 0.0001
@@ -83,8 +90,8 @@ n_step = n_epoch * n_step_epoch
 shuffle_buffer_size = 128
 
 optimizer = tf.optimizers.Adam(learning_rate)
-# optimizer = tf.optimizers.SGD(learning_rate)
 cost = tl.cost.cross_entropy
+
 
 def generator_train():
     inputs = X_train
@@ -131,7 +138,6 @@ def _map_fn_test(img, target):
     return img, target
 
 
-
 def _train_step(network, X_batch, y_batch, cost, train_op=tf.optimizers.Adam(learning_rate=0.0001), acc=None):
     with tf.GradientTape() as tape:
         y_pred = network(X_batch)
@@ -144,8 +150,10 @@ def _train_step(network, X_batch, y_batch, cost, train_op=tf.optimizers.Adam(lea
     else:
         return _loss, None
 
+
 def accuracy(_logits, y_batch):
     return np.mean(np.equal(np.argmax(_logits, 1), y_batch))
+
 
 # dataset API and augmentation
 train_ds = tf.data.Dataset.from_generator(
@@ -179,6 +187,10 @@ for epoch in range(n_epoch):
         train_loss += _loss
         train_acc += acc
         n_iter += 1
+
+        print("Epoch {} of {} took {}".format(epoch + 1, n_epoch, time.time() - start_time))
+        print("   train loss: {}".format(train_loss / n_iter))
+        print("   train acc:  {}".format(train_acc / n_iter))
 
     # use training and evaluation sets to evaluate the model every print_freq epoch
     if epoch + 1 == 1 or (epoch + 1) % print_freq == 0:
